+++ changelog.md
@@ -10,6 +10,7 @@ version 5.4.0 - upcoming
   * added colors/icons to request types and status information (can be disabled in options)
   * disabled javafx off-color subpixel antialiasing for font (macOs)
   * openapi import: importing server-urls as environment variables
+  * added colors to environments
 
 version 5.3.0
   * mask plain-keys on overview